+++ Common/Common.csproj
@@ -117,6 +117,7 @@
     <Compile Include="Helpers\DynamicValuesHelper.cs" />
     <Compile Include="Helpers\FunctionalHelper.cs" />
     <Compile Include="Utility\DocDbResourceType.cs" />
+    <Compile Include="Utility\DocDbResourceTypeHelper.cs" />
     <Compile Include="Utility\DocDbRestUtility.cs" />
     <Compile Include="Utility\IDocDbRestUtility.cs" />
     <Compile Include="Helpers\ParsingHelper.cs" />
